Refactor tagging, deletion, and interval logic
TAGS: Autosnap will now tag the snapshots it makes with the same
user-configured tag specified in the config file. It will also use that
tag to filter out which snapshots to delete and not delete.

DELETION: Autosnap now bases its deletion decisions based on the NUMBER
of snapshots, not the AGE of the snapshots.  This allows two things to happen:
    1. This can be run in an on-demand basis instead of having to be run
      on a schedule.
    2. This prevents autosnap from deleting all of the snapshots for a volume
      if it hasn't backed up in more than X number of days. For instance, if
      the script fails to run over a weekend and you only want to keep the
      last 3 days, it would previously have deleted all the snapshots if
      you came back on Monday and manually ran the script (or fixed the
      issue).

INTERVAL: This commit also removes the concept of 'daily', 'weekly' or
'monthly' snapshots. The new setting only configures how many snapshots
you want to keep, period. No more having different sets of snapshots to
keep track of.