Refine selection of snapshots to delete
Now only selects a snapshot if it has both the tag and value present.
Previously, it would only see if 'autosnap' was present in any of the
tag values. That could lead to unintended selection if another tag had
the string in it (even though it wasn't OUR 'snapshot_type' tag).

Also, this now filters the snapshots directly from the AWS API instead
of pulling all the tags down and filtering it locally.